restructure src directory and include package-lock changes
