+++ src/components/GraphPanel.js
