+++ src/index.js
@@ -1,6 +1,6 @@
 import React from 'react'
 import { render } from 'react-dom'
-import App from './App'
+import App from './components/App'
 
 const root = document.createElement('div')
 document.body.appendChild(root)
